detect-ssl-state: move unittests to tests/
